// src/interfaces/IPersonTvCast.ts
export interface IPersonTvCast {
  credit_id: string;
  original_name: string;
  id: number;
  genre_ids: number[];
  character: string;
  name: string;
  poster_path: string | null;
  vote_count: number;
  vote_average: number;
  popularity: number;
  episode_count: number;
  original_language: string;
  first_air_date: string;
  backdrop_path: string | null;
  overview: string;
  origin_country: string[];
}